Revert "refactor to prevent bugs - select pair inside pairselector, not trade page"
This reverts commit c877fd41e766535cc09ffbb7313605bcf6453822.
